CRUD on Movie

+++ Services/Function/Movie/Commands/DeleteMovie/DeleteMovieHandler.cs
@@ -4,9 +4,9 @@ using AutoMapper;
 using Domain.Repositories;
 using MediatR;
 
-namespace Services.Function.Movie.Commands
+namespace Services.Function.Movie.Commands.DeleteMovie
 {
-    public class DeleteMovieHandler : IRequestHandler<DeleteMovieCommand>
+    public class DeleteMovieHandler : IRequestHandler<DeleteMovieCommand, BaseResponse>
     {
         private readonly IMovieAsyncRepository _movieAsyncRepository;
         private readonly IMapper _mapper;
@@ -17,13 +17,18 @@ namespace Services.Function.Movie.Commands
             _mapper = mapper;
         }
 
-        public async Task<Unit> Handle(DeleteMovieCommand request, CancellationToken cancellationToken)
+        public async Task<BaseResponse> Handle(DeleteMovieCommand request, CancellationToken cancellationToken)
         {
             var movieToDelete = await _movieAsyncRepository.GetByIdAsync(request.MovieId);
 
+            if (movieToDelete is null)
+            {
+                return new BaseResponse($"Movie with id {request.MovieId} not found", false);
+            }
+
             await _movieAsyncRepository.DeleteAsync(movieToDelete);
 
-            return Unit.Value;
+            return new BaseResponse();
         }
 
     }
